EKS/Strands/requirements.txt
# AWS AgentCore and Strands dependencies
bedrock-agentcore>=0.1.21
bedrock-agentcore-starter-toolkit>=0.1.21
strands-agents>=1.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# HTTP requests for weather API
requests>=2.31.0

# Optional but recommended for production
python-dotenv>=1.0.0

# Type hints support (if needed)
typing-extensions>=4.9.0